remove current triggers, start refactoring them to the new format
