--- src/egamma_tnp/triggers/__init__.py
@@ -1,17 +1,0 @@
-from egamma_tnp.triggers.doubleelept_caloidl_mw import DoubleElePt_CaloIdL_MW
-from egamma_tnp.triggers.elept1_elept2_caloidl_trackidl_isovl import (
-    ElePt1_ElePt2_CaloIdL_TrackIdL_IsoVL,
-)
-from egamma_tnp.triggers.elept_caloidvt_gsftrkidt import ElePt_CaloIdVT_GsfTrkIdT
-from egamma_tnp.triggers.elept_wptight_gsf import ElePt_WPTight_Gsf
-
-__all__ = (
-    "DoubleElePt_CaloIdL_MW",
-    "ElePt_WPTight_Gsf",
-    "ElePt_CaloIdVT_GsfTrkIdT",
-    "ElePt1_ElePt2_CaloIdL_TrackIdL_IsoVL",
-)
-
-
-def dir():
-    return __all__
