--- src/egamma_tnp/triggers/basesingleelectrontrigger.py
@@ -1,238 +1,0 @@
-import os
-from functools import partial
-
-from coffea.dataset_tools import apply_to_fileset
-from coffea.nanoevents import NanoAODSchema
-
-
-class BaseSingleElectronTrigger:
-    """BaseSingleElectronTrigger class for HLT Trigger efficiency from NanoAOD.
-
-    This class holds the basic methods for all the Tag and Probe classes for different single electron triggers.
-    """
-
-    def __init__(
-        self,
-        fileset,
-        tnpimpl_class,
-        pt,
-        filterbit,
-        avoid_ecal_transition_tags,
-        avoid_ecal_transition_probes,
-        goldenjson,
-        extra_filter,
-        extra_filter_args,
-    ):
-        self.fileset = fileset
-        self._tnpimpl_class = tnpimpl_class
-        self.pt = pt
-        self.filterbit = filterbit
-        self.avoid_ecal_transition_tags = avoid_ecal_transition_tags
-        self.avoid_ecal_transition_probes = avoid_ecal_transition_probes
-        self.goldenjson = goldenjson
-        self._extra_filter = extra_filter
-        self._extra_filter_args = extra_filter_args
-
-        if goldenjson is not None and not os.path.exists(goldenjson):
-            raise FileNotFoundError(f"Golden JSON {goldenjson} does not exist.")
-
-    def get_tnp_arrays(
-        self,
-        schemaclass=NanoAODSchema,
-        uproot_options=None,
-        compute=False,
-        scheduler=None,
-        progress=False,
-    ):
-        """Get the Pt, Eta and Phi arrays of the passing and all probes.
-        WARNING: Not recommended to be used for large datasets as the arrays can be very large.
-
-        Parameters
-        ----------
-            schemaclass: BaseSchema, default NanoAODSchema
-                The nanoevents schema to interpret the input dataset with.
-            uproot_options : dict, optional
-                Options to pass to uproot. Pass at least {"allow_read_errors_with_report": True} to turn on file access reports.
-            compute : bool, optional
-                Whether to return the computed arrays or the delayed arrays.
-                The default is False.
-            scheduler : str, optional
-                The dask scheduler to use. The default is None.
-                Only used if compute is True.
-            progress : bool, optional
-                Whether to show a progress bar if `compute` is True. The default is False.
-                Only meaningful if compute is True and no distributed Client is used.
-
-        Returns
-        -------
-            arrays :a tuple of dask awkward zip items of the form (passing_probes, all_probes).
-                Each of the zip items has the following fields:
-                    pt: dask_awkward.Array
-                        The Pt array of the probes.
-                    eta: dask_awkward.array
-                        The Eta array of the probes.
-                    phi: dask_awkward.array
-                        The Phi array of the probes.
-            report: dict of awkward arrays of the same form as fileset.
-                For each dataset an awkward array that contains information about the file access is present.
-        """
-        if uproot_options is None:
-            uproot_options = {}
-
-        perform_tnp = self._tnpimpl_class(
-            pt=self.pt,
-            filterbit=self.filterbit,
-            avoid_ecal_transition_tags=self.avoid_ecal_transition_tags,
-            avoid_ecal_transition_probes=self.avoid_ecal_transition_probes,
-            goldenjson=self.goldenjson,
-            extra_filter=self._extra_filter,
-            extra_filter_args=self._extra_filter_args,
-        )
-        data_manipulation = perform_tnp
-
-        to_compute = apply_to_fileset(
-            data_manipulation=data_manipulation,
-            fileset=self.fileset,
-            schemaclass=schemaclass,
-            uproot_options=uproot_options,
-        )
-        if compute:
-            import dask
-            from dask.diagnostics import ProgressBar
-
-            if progress:
-                pbar = ProgressBar()
-                pbar.register()
-
-            computed = dask.compute(to_compute, scheduler=scheduler)
-
-            if progress:
-                pbar.unregister()
-
-            return computed[0]
-
-        return to_compute
-
-    def get_tnp_histograms(
-        self,
-        schemaclass=NanoAODSchema,
-        uproot_options=None,
-        plateau_cut=None,
-        eta_regions_pt=None,
-        eta_regions_eta=None,
-        eta_regions_phi=None,
-        compute=False,
-        scheduler=None,
-        progress=False,
-    ):
-        """Get the Pt, Eta and Phi histograms of the passing and all probes.
-
-        Parameters
-        ----------
-            schemaclass: BaseSchema, default NanoAODSchema
-                The nanoevents schema to interpret the input dataset with.
-            uproot_options : dict, optional
-                Options to pass to uproot. Pass at least {"allow_read_errors_with_report": True} to turn on file access reports.
-            plateau_cut : int or float, optional
-                The Pt threshold to use to ensure that we are on the efficiency plateau for eta and phi histograms.
-                The default None, meaning that no extra cut is applied and the activation region is included in those histograms.
-            eta_regions_pt : dict, optional
-                A dictionary of the form `{"name": [etamin, etamax], ...}`
-                where name is the name of the region and etamin and etamax are the absolute eta bounds.
-                The Pt histograms will be split into those eta regions.
-                The default is to avoid the ECAL transition region meaning |eta| < 1.4442 or 1.566 < |eta| < 2.5.
-            eta_regions_eta : dict, optional
-                A dictionary of the form `{"name": [etamin, etamax], ...}`
-                where name is the name of the region and etamin and etamax are the absolute eta bounds.
-                The Eta histograms will be split into those eta regions.
-                The default is to use the entire |eta| < 2.5 region.
-            eta_regions_phi : dict, optional
-                A dictionary of the form `{"name": [etamin, etamax], ...}`
-                where name is the name of the region and etamin and etamax are the absolute eta bounds.
-                The Phi histograms will be split into those eta regions.
-                The default is to use the entire |eta| < 2.5 region.
-            compute : bool, optional
-                Whether to return the computed hist.Hist histograms or the delayed hist.dask.Hist histograms.
-                The default is False.
-            scheduler : str, optional
-                The dask scheduler to use. The default is None.
-                Only used if compute is True.
-            progress : bool, optional
-                Whether to show a progress bar if `compute` is True. The default is False.
-                Only meaningful if compute is True and no distributed Client is used.
-
-        Returns
-        -------
-            A tuple of the form (histograms, report) if `allow_read_errors_with_report` is True, otherwise just histograms.
-            histograms : dict of dicts of the same form as fileset where for each dataset the following dictionary is present:
-                A dictionary of the form `{"var": {"name": {"passing": passing_probes, "all": all_probes}, ...}, ...}`
-                where `"var"` can be `"pt"`, `"eta"`, or `"phi"`.
-                Each `"name"` is the name of eta region specified by the user and `passing_probes` and `all_probes` are `hist.dask.Hist` objects.
-                These are the histograms of the passing and all probes respectively.
-            report: dict of awkward arrays of the same form as fileset.
-                For each dataset an awkward array that contains information about the file access is present.
-        """
-        if uproot_options is None:
-            uproot_options = {}
-
-        perform_tnp = self._tnpimpl_class(
-            pt=self.pt,
-            filterbit=self.filterbit,
-            avoid_ecal_transition_tags=self.avoid_ecal_transition_tags,
-            avoid_ecal_transition_probes=self.avoid_ecal_transition_probes,
-            goldenjson=self.goldenjson,
-            extra_filter=self._extra_filter,
-            extra_filter_args=self._extra_filter_args,
-        )
-        data_manipulation = partial(
-            self._make_tnp_histograms,
-            perform_tnp=perform_tnp,
-            plateau_cut=plateau_cut,
-            eta_regions_pt=eta_regions_pt,
-            eta_regions_eta=eta_regions_eta,
-            eta_regions_phi=eta_regions_phi,
-        )
-
-        to_compute = apply_to_fileset(
-            data_manipulation=data_manipulation,
-            fileset=self.fileset,
-            schemaclass=schemaclass,
-            uproot_options=uproot_options,
-        )
-        if compute:
-            import dask
-            from dask.diagnostics import ProgressBar
-
-            if progress:
-                pbar = ProgressBar()
-                pbar.register()
-
-            computed = dask.compute(to_compute, scheduler=scheduler)
-
-            if progress:
-                pbar.unregister()
-
-            return computed[0]
-
-        return to_compute
-
-    def _make_tnp_histograms(
-        self,
-        events,
-        perform_tnp,
-        plateau_cut,
-        eta_regions_pt,
-        eta_regions_eta,
-        eta_regions_phi,
-    ):
-        from egamma_tnp.utils import fill_cutncount_histograms
-
-        passing_probes, all_probes = perform_tnp(events)
-        return fill_cutncount_histograms(
-            passing_probes,
-            all_probes,
-            plateau_cut=plateau_cut,
-            eta_regions_pt=eta_regions_pt,
-            eta_regions_eta=eta_regions_eta,
-            eta_regions_phi=eta_regions_phi,
-        )
